Reimplementation of the turning smart screen python.
This is a reimplementation from scratch, using what was learnt in
the GPL library, but made into an object based implementation with
hopefully support for both devices.

This is a re-write, released under MIT license, as the project I
wish to use it in is not under GPL license and I do not wish to
convert it to be so.