Optimise the delay between bitmap and command.
The delay between the bitmap and the command was forced to be 0.01s
which meant that the Python code couldn't do anything whilst it was
waiting for the display to do its work.

This change removes the forced sleep of 0.01s and instead lets the
code run and only sleeps if a command comes within the inter-bitmap
delay. This means that we parallelise better. The delay we have to
allow is longer than 0.01s (0.02s is used). This implies that the
delay we had was actaully too short and we were only working because
the processing between the bitmap and the next command was taking
up the slack time. If that is true then there would have been
instances where machines were faster and the result was a corrupt
image, but on slower machine there would be no problem.