Add support for paletted Pillow images; fix out of bounds image writes.
Paletted Pillow images are now supported. The image bounds could be
exceeded when there was a larger image written than the expected. We
now truncate the bounds at the limits of the protocol for the variant
1 displays.

I've still been unable to get the variant 1 display to function
with my code (or possibly the other way around). This might be a macOS
problem or something else.